requirements.txt
asgiref==3.5.2
django==4.0.6
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
django-environ==0.9.0
pillow==9.2.0
coverage==6.4.2
djoser
djangorestframework-simplejwt
pyjwt
django-countries==7.3.2
flake8==5.0.4
black==22.6.0
isort==5.10.1
redis==4.3.4
celery==5.2.7
flower==1.2.0
django-celery-email==3.0.0
django-filter==22.1
celery-error-emails==0.3